refactor: using socket channel
