+++ packages/connection/src/common/index.ts
@@ -2,6 +2,6 @@ export * from './message';
 export * from './proxy';
 export * from './rpcProtocol';
 export * from './utils';
-export * from './ws-channel';
+export * from './socket-channel';
 export * from './connect';
 export * from './types';
