+++ packages/connection/src/node/common-channel-handler.ts
@@ -3,8 +3,8 @@ import WebSocket from 'ws';
 
 import { PlatformBuffer } from '@opensumi/ide-core-common/lib/connection/types';
 
+import { SocketChannel, ChannelMessage } from '../common/socket-channel';
 import { stringify, parse } from '../common/utils';
-import { WSChannel, ChannelMessage } from '../common/ws-channel';
 
 import { WebSocketHandler, CommonChannelHandlerOptions } from './ws';
 
@@ -32,7 +32,7 @@ export class CommonChannelPathHandler {
     }
     const handlerArr = this.handlerMap.get(channelToken) as IPathHandler[];
     const handlerFn = handler.handler.bind(handler);
-    const setHandler = (connection: WSChannel, clientId: string, params) => {
+    const setHandler = (connection: SocketChannel, clientId: string, params) => {
       handler.connection = connection;
       handlerFn(connection, clientId, params);
     };
@@ -88,7 +88,7 @@ export class CommonChannelHandler extends WebSocketHandler {
   public handlerId = 'common-channel';
   private wsServer: WebSocket.Server;
   protected handlerRoute: MatchFunction;
-  private channelMap: Map<string, WSChannel> = new Map();
+  private channelMap: Map<string, SocketChannel> = new Map();
   private connectionMap: Map<string, WebSocket> = new Map();
   private heartbeatMap: Map<string, NodeJS.Timeout> = new Map();
 
@@ -140,7 +140,7 @@ export class CommonChannelHandler extends WebSocketHandler {
 
             // 生成 channel 对象
             const connectionSend = this.channelConnectionSend(connection);
-            const channel = new WSChannel(connectionSend, channelId);
+            const channel = new SocketChannel(connectionSend, channelId);
             this.channelMap.set(channelId, channel);
 
             // 根据 path 拿到注册的 handler
